requirements.txt
requests
pygithub
tornado
SQLAlchemy